td-cloud/requirements.txt
flask==3.0.3
redis==5.0.8
gunicorn==22.0.0
flask-socketio==5.4.1
eventlet==0.37.0
google-cloud-pubsub==2.27.1
simple-websocket==1.0.0
